+++ src/jobs/RebuildDailyStatsJob.php
@@ -6,6 +6,7 @@ use Craft;
 use craft\base\Batchable;
 use craft\queue\BaseBatchedJob;
 use fostercommerce\bestsellers\Plugin;
+use Throwable;
 
 class RebuildDailyStatsJob extends BaseBatchedJob
 {
@@ -28,8 +29,14 @@ class RebuildDailyStatsJob extends BaseBatchedJob
 	{
 		/** @var string $date */
 		$date = $item;
-		$plugin = Plugin::getInstance();
-		$plugin->dailyStats->aggregateDay($date);
+
+		try {
+			$plugin = Plugin::getInstance();
+			$plugin->dailyStats->aggregateDay($date);
+		} catch (Throwable $throwable) {
+			Craft::warning("Failed to aggregate daily stats for {$date}: {$throwable->getMessage()}", 'best-sellers');
+			Plugin::getInstance()->backfillLogs->log('daily-stats', $date, $throwable->getMessage());
+		}
 	}
 
 	protected function defaultDescription(): ?string